Open Issues Completed
Error Messages now OrangeRed changed from Red to make easier to see in
daylight on the black background.
Added some code to size the controls in case they are moved during
development.
Added a Parameters window and a button to switch on and off. The window
is not docked so you can move it around and tab to it should you be
reviewing it against the current analysis.