Pass Command Line Arguments to UpdaterProgram
Changed the way in which the parameters are passed to the
UpdaterProgram. We now do this using Command Line Arguments, we have
removed the need to create a file called update.txt.
Also, added a Primary and Secondary update server URL's, these are
checked and the best one passed to the UpdaterProgram as a Command Line
Argument.