Working the best
This is the best solution so far.

' In testing my files I found that a motor produces a minimum of 60%
' A non-motor produces a maximum of 12% where the value was 1500 PWM 99%
of the time,
' and 8% where the value was 1100 PWM 99% of the time.
' Therefore the MotorDetectionThreshold is currently set at 40%

From this point I will work on creating a different version to see if I
can get better results based on channel fluctuations rather then channel
averages.