Based on Fluctuations
' In the end we may need to make a mixture of the Channel Average code
below and this code.
' The Channel Average code creates higher numbers for the motor channels
but has the
' significant draw back that when a non-channel remains within the scan
limits it is counted as a motor.
' It is for this reason the fluctuation code was created. However, this
code may fail if a gimbal was being
' driven by the Flight Controller because the fluctuations on those
channels may be within the Motor range.
' For this reason we must keep the MotorDetectionMinFluctuation value as
high as possible and the
' MotorDetectionThreshold as low as possible.
' The other trouble with this code is that when a quad is flying forward
the front motors do much less than
' then side and back motors. So for a Hex motors 1 and 2 usually produce
much lower results than all the other motors.
' To speed up the code we monitor the % of Channels 1,2 and 3, if they
fall below MotorDetectionEnd then we stop. The
' reason for this is that channels will find the first calcualtion as
50%, and will start to drop quickly. Only
' motor channels seem to do this.
' Also to speed up the code we monitor the number of succcessful hits on
channels 1,2, and 3, once these reach
' MotorDetectionEnd_Count then we also stop.